feat: finished project sneakers
