+++ src/components/Drawer/drawer.module.scss
@@ -1,0 +1,48 @@
+.overlay {
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  z-index: 1;
+  visibility: hidden;
+  overflow: hidden;
+  opacity: 0;
+  transition: opacity 0.1s ease-in-out, visibility 0.1s ease-in-out;
+}
+
+.overlayVisible {
+  opacity: 1;
+  visibility: visible;
+  .drawer {
+    transform: translateX(0);
+  }
+}
+
+.drawer {
+  display: flex;
+  flex-direction: column;
+  position: absolute;
+  width: 420px;
+  height: 100%;
+  right: 0;
+  background: #ffffff;
+  box-shadow: -10px 4px 24px 0px #0000001a;
+  padding: 30px;
+  transform: translateX(100%);
+  transition: transform 0.3s ease-in-out;
+
+  h2 {
+    margin: 0;
+    font-size: 24px;
+  }
+  .items {
+    overflow: auto;
+    margin-bottom: 40;
+    flex: 1;
+    flex-direction: column;
+    justify-content: space-between;
+  }
+  z-index: 2;
+}
