+++ src/index.scss
@@ -75,38 +75,8 @@ ul.d-flex {
   }
 }
 
-.overlay {
-  position: absolute;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  z-index: 1;
-}
 
-.drawer {
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-  width: 420px;
-  height: 100%;
-  right: 0;
-  background: #ffffff;
-  box-shadow: -10px 4px 24px 0px #0000001a;
-  padding: 30px;
-  h2 {
-    margin: 0;
-    font-size: 24px;
-  }
-  .items {
-    overflow: auto;
-    margin-bottom: 40;
-    flex: 1;
-    flex-direction: column;
-    justify-content: space-between;
-  }
-}
+
 
 .emptyCart {
   display: flex;
